+++ src/Components/Scenes/Hydration/NewJob/NewJob.js
@@ -1,0 +1,39 @@
+import React from 'react';
+import { Box } from '@material-ui/core';
+import { makeStyles } from '@material-ui/core/styles';
+import NewJobHeader from './NewJobHeader';
+
+const chromeStyles = makeStyles(theme => ({
+  container: {
+    position: 'fixed',
+    zIndex: 1101,
+    height: '100%',
+    width: '100%',
+    display: 'flex'
+  },
+  canvas: {
+    background:
+      'repeating-linear-gradient(45deg, #EDF5FF, #EDF5FF 15px, #FFFFFF 5px, #FFFFFF 20px)',
+    height: 'inherit',
+    flexGrow: 1
+  },
+  toolbar: {
+    minWidth: '350px',
+    backgroundColor: 'white'
+  }
+}));
+
+const HydrationChrome = props => {
+  const classes = chromeStyles();
+
+  return (
+    <Box className={classes.container}>
+      <Box className={classes.canvas}>
+        <NewJobHeader jobName="My New Job"></NewJobHeader>
+      </Box>
+      <Box className={classes.toolbar}>Toolbar Placeholder</Box>
+    </Box>
+  );
+};
+
+export default HydrationChrome;
